refactor: move annotations to sep file
